backend/requirements.txt
fastapi
uvicorn
pandas
python-dotenv
python-multipart
numpy
scikit-learn
openai
